Function futures_spsc::unbounded

pub fn unbounded<T>() -> (UnboundedSender<T>, UnboundedReceiver<T>)

Creates an unbounded sender/receiver pair.

This function creates a new single-producer, single-consumer (spsc) channel where the sender cannot be cloned and all messages are buffered internally. That is, this channel provides no backpressure as it will buffer items unboundedly. For a bounded and backpressure-based solution, see the channel function.

An unbounded channel's sender is always ready to send a message and all message sends succeed so long as the receiver has not gone away yet. Note that successfully sending a message does not imply receiving the message as the receiver is not guaranteed to receive all messages before being dropped.

Also note that the lack of backpressure here can be a risky choice for some situations and may lead to situations such as resource exhaustion if a system is overloaded. Care should be taken to ensure there are limits on the system elsewhere to prevent this case.